Chore Bot project
Really proud of myself with this one. My first HTML/CSS/javaScript combo project.
Shure it's a simple game, and in the bulk of the lesson there's alot of handholding. But i felt comfortable doing it and only used the 'hints' to check that my work was correct.

The score counter is selfimplemented no instructions and no googling!